app/requirements.txt
discord.py==2.0.1
canvasapi==3.0.0
pygsheets==2.0.5
beautifulsoup4==4.11.1